Create Styles: Improve ThemeProvider initial (theme) style rendering (#53)
This update improves the ThemeProvider initial (theme) style rendering. The
change needed to achieve this involved coordinating the injectGlobal function
call in the ThemeProvider generated from the createStyleSystem factory.

The initial injectGlobal is the same as a coreElement, using the
useHydrateGlobalStyles hook.

Ultimately, this helps with SSR rendering for custom theme styles.