Components: Add ListGroup component
This update adds a new ListGroup component, which is used to group clusters of controls, which are often rendered in either a Sidebar or Toolbar context.
ListGroup accounts for certain children content, such as Grid, and makes adjustments to the props to ensure consistent gap spacing.
This is achieved using the new "shallow" ComponentsProvider rendering property.

"Shallow" props only apply to the first level (regardless of nested position). They do no propogate down.

In the context of ListGroup, it ensures that only the direct child `Grid` is affected. Nested `Grid` components will not be adjusted.